DDP-5402 - refactored: renamed module package to 'activityinstancebuilder' and main class to 'ActivityInstanceFromActivityDefStoreBuilder'; fixed templates detection from 'ActivityDefStore'; add a method which starts ActivityInstance building to 'ActivityInstanceService'.
