+++ pepper-apis/src/main/java/org/broadinstitute/ddp/service/actvityinstancebuilder/block/question/impl/AgreementQuestionCreator.java
@@ -1,0 +1,35 @@
+package org.broadinstitute.ddp.service.actvityinstancebuilder.block.question.impl;
+
+import org.broadinstitute.ddp.model.activity.definition.question.AgreementQuestionDef;
+import org.broadinstitute.ddp.model.activity.instance.answer.AgreementAnswer;
+import org.broadinstitute.ddp.model.activity.instance.question.AgreementQuestion;
+import org.broadinstitute.ddp.service.actvityinstancebuilder.ActivityInstanceFromActivityDefStoreBuilder;
+import org.broadinstitute.ddp.service.actvityinstancebuilder.ElementCreator;
+import org.broadinstitute.ddp.service.actvityinstancebuilder.block.question.QuestionCreator;
+
+import static org.broadinstitute.ddp.service.actvityinstancebuilder.block.question.QuestionUtil.isReadOnly;
+
+/**
+ * Creates {@link AgreementQuestion}
+ */
+public class AgreementQuestionCreator extends ElementCreator {
+
+    public AgreementQuestionCreator(ActivityInstanceFromActivityDefStoreBuilder.Context context) {
+        super(context);
+    }
+
+    public AgreementQuestion createAgreementQuestion(QuestionCreator questionCreator, AgreementQuestionDef questionDef) {
+        return new AgreementQuestion(
+                questionDef.getStableId(),
+                getTemplateId(questionDef.getPromptTemplate()),
+                questionDef.isRestricted(),
+                questionDef.isDeprecated(),
+                isReadOnly(context, questionDef),
+                getTemplateId(questionDef.getTooltipTemplate()),
+                getTemplateId(questionDef.getAdditionalInfoHeaderTemplate()),
+                getTemplateId(questionDef.getAdditionalInfoFooterTemplate()),
+                questionCreator.getAnswers(AgreementAnswer.class, questionDef.getStableId()),
+                questionCreator.getValidationRules(questionDef)
+        );
+    }
+}
